requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.1.1
typing_extensions==4.14.0
Werkzeug==3.1.3
WTForms==3.2.1
PyMySQL==1.1.1